Playing the sound of words through Control
